refactor(donor-research): extract candidate-card reason logic; a11y + suppress best-effort catch
